Added statmechTest.py
With test_rotor_symmetry_determination and a relocated
test_gaussian_log_file_error
Removed the TestStatmech class from commonTest.py
